requirements.txt
Pillow>=10.3.0,<10.4
django>=5.0.2,<5.1
djangorestframework>=3.15.0,<3.16
psycopg2-binary>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
django-cors-headers>=4.3.1,<4.4
uwsgi>=2.0.20,<2.1
PyJWT>=2.8.0,<2.9
python-keycloak>=4.3.0,<4.4
channels>=4.1.0,<4.2
channels-redis>=4.2.0,<4.3
python-dotenv>=1.0.1,<1.1